requirements.txt
azure.storage.blob
pandas
openai
python-dotenv
pdfplumber
fitz
flask